TEST: Moved test_model_time_step_control
