+++ tests/numerics/test_time_step_control.py
@@ -777,229 +777,3 @@ class TestTimeControl:
         )
         # Remove temporary file.
         pth.unlink()
-
-
-class DynamicTimeStepTestCaseModel(SinglePhaseFlow):
-    """A mockup model that overrides `check_convergence` and predefines convergence
-    behavior after each nonlinear iteration.
-
-    See the description of the input parameters at `test_model_time_step_control`.
-
-    """
-
-    def __init__(
-        self,
-        num_nonlinear_iterations: list[int],
-        time_step_converged: list,
-        params: dict,
-    ):
-        super().__init__(params)
-        self.time_step_idx: int = -1
-        self.num_nonlinear_iters: int = 0
-        self.num_nonlinear_iterations: list[int] = num_nonlinear_iterations
-        self.time_step_converged: list = time_step_converged
-        self.time_step_history: list = []
-
-    def before_nonlinear_loop(self) -> None:
-        super().before_nonlinear_loop()  # The AD time step is expected to update here.
-        self.time_step_idx += 1
-        self.num_nonlinear_iters = 0
-        self.time_step_history.append(self.time_manager.dt)
-
-    def before_nonlinear_iteration(self):
-        super().before_nonlinear_iteration()
-
-        # The AD time step should not change throughout the Newton iterations.
-        assert (
-            self.equation_system.evaluate(self.ad_time_step) == self.time_manager.dt
-        ), "The AD time step value conflicts with the value from the time_manager."
-
-        # The initial guess for the unknown time step values should be equal to the
-        # known time step values. See https://github.com/pmgbergen/porepy/issues/1205.
-        if self.num_nonlinear_iters == 0:
-            iterate_values = self.equation_system.get_variable_values(iterate_index=0)
-            state_values = self.equation_system.get_variable_values(time_step_index=0)
-            assert np.all(iterate_values == state_values), (
-                "Likely, 'iterate' was not reset after the unsuccessful time step."
-            )
-
-        self.num_nonlinear_iters += 1
-
-    def _is_nonlinear_problem(self):
-        return True
-
-    # Minimizing computational expenses.
-    def assemble_linear_system(self) -> None:
-        pass
-
-    def solve_linear_system(self) -> np.ndarray:
-        return np.ones(self.equation_system.num_dofs())
-
-MAX_NONLINEAR_ITER = 10
-
-
-@pytest.mark.parametrize(
-    "params",
-    [
-        # Case 1: A successful simulation run with dynamic time stepping.
-        # Covers these situations:
-        # - decrease the time step after diverged
-        # - decrease the time step after iteration limit
-        # - increase the time step due to few nonlinear iterations
-        # - keep the time step due to expected number of nonlinear iterations
-        # - decrease the time step due to many nonlinear iterations (after convergence)
-        # - decrease the time step to meet the schedule (last time step)
-        {
-            # Below reads as: time step 0 takes 4 nonlinear iterations, time step 1
-            # takes 3 nonlinear iterations, etc.
-            "num_nonlinear_iterations": [4, 3, MAX_NONLINEAR_ITER + 2, 1, 6, 9, 1, 1],
-            # Time step 0 diverged after 4 iterations, time step 1 converged after 3
-            # iterations, etc. "unreachable" means that the convergence check should not
-            # be called due to exceeding the iteration limit.
-            "time_step_converged": [False, True, "unreachable"] + [True] * 5,
-            # Time step magnitudes to compare with. These are known values produced with
-            # the settings of the TimeStepper found in the test function below.
-            "exported_dt_expected": [1, 0.3, 0.6, 0.18, 0.36, 0.36, 0.144, 0.006],
-        },
-        # Case 2: constant_dt. Should fail after nonlinear divergence.
-        {
-            "constant_dt": True,
-            "num_nonlinear_iterations": [2, 3],
-            "time_step_converged": [True, False],
-            "exported_dt_expected": [1, 1],
-            "schedule_end": 2,  # Matches the constant dt = 1.
-            "failure_reason": "Max retries (1)",
-        },
-        # Case 3: An unsuccessful simulation with dynamic time stepping. Reached the
-        # minimal time step and should fail.
-        {
-            "num_nonlinear_iterations": [1, 1, 1],
-            "time_step_converged": [False, False, False],
-            "exported_dt_expected": [1, 0.3, 0.1],
-            "failure_reason": "time step achieved its minimum admissible value",
-        },
-        # Case 4: The time step fails right before the schedule point. Expected to
-        # decrease dt and meet the schedule regardless.
-        {
-            "num_nonlinear_iterations": [1, 1, 1, 1, 1],
-            "time_step_converged": [True, False, True, True, True],
-            "exported_dt_expected": [1, 0.35, 0.105, 0.21, 0.035],
-        },
-        # Case 5: Fail because we reach the limit of attempts to cut the time step.
-        {
-            "num_nonlinear_iterations": [1, 1, 1, 1, 1, 1],
-            "time_step_converged": [True, True, False, False, False, False],
-            "exported_dt_expected": [1, 2, 4, 1.2, 0.36, 0.108],
-            "schedule_end": 10,  # Far beyond possible dt to avoid dt_max clipping.
-            "failure_reason": "Max retries (4)",
-        },
-        # Case 6: All time steps are successful so dt reaches dt_max and remains it.
-        {
-            "num_nonlinear_iterations": [1, 1, 1, 1, 1, 1],
-            "time_step_converged": [True, True, True, True, True, True],
-            "exported_dt_expected": [1, 2, 4, 5, 5, 0.1],
-            "schedule_end": 17.1,
-        },
-        # Case 7: All time steps are successful, but take too many nonlinear iterations,
-        # so dt reaches dt_min and remains it.
-        {
-            "num_nonlinear_iterations": [8, 8, 8, 8, 8, 8],
-            "time_step_converged": [True, True, True, True, True, True],
-            "exported_dt_expected": [1, 0.4, 0.16, 0.1, 0.1, 0.1],
-            "schedule_end": 1.86,
-        },
-        # Case 8: Time step we need to make to meet the schedule is below minimal dt.
-        # The current behavior is that we ignore dt_min to meet the schedule.
-        # This implies that if we hit this in the middle of the simulation, next dt will
-        # be dt_min, which is quite sub-optimal.
-        {
-            "num_nonlinear_iterations": [1, 1],
-            "time_step_converged": [True, True],
-            "exported_dt_expected": [1, 1e-5],
-            "schedule_end": 1 + 1e-5,
-        },
-    ],
-)
-def test_model_time_step_control(params: dict):
-    """The integration test of the `TimeManager` class into PorePy models."""
-    constant_dt = params.get("constant_dt", False)
-    num_nonlinear_iterations = params["num_nonlinear_iterations"]
-    time_step_converged = params["time_step_converged"]
-    exported_dt_expected = params["exported_dt_expected"]
-    failure_reason = params.get("failure_reason", "")
-    # 1.35 is the default value assumed by most of the tests. It is arbitrary.
-    schedule_end = params.get("schedule_end", 1.35)
-
-    should_fail = len(failure_reason) != 0
-
-    time_manager = pp.TimeManager(
-        schedule=(0, schedule_end),
-        dt_init=1,
-        constant_dt=constant_dt,
-        dt_min_max=(0.1, 5),
-        iter_relax_factors=(0.4, 2),
-        iter_optimal_range=(4, 7),
-        recomp_factor=0.3,
-        recomp_max=4,
-    )
-
-    class DynamicNewtonSolver(pp.NewtonSolver):
-        def check_convergence(
-            self, model: DynamicTimeStepTestCaseModel, nonlinear_increment
-        ) -> tuple[
-            ConvergenceStatusCollection,
-            ConvergenceStatusCollection,
-            ConvergenceInfoCollection,
-        ]:
-            if (
-                model.nonlinear_solver_statistics.num_iterations
-                < model.num_nonlinear_iterations[model.time_step_idx] - 1
-            ):
-                return (
-                    ConvergenceStatusCollection(
-                        {"crit": ConvergenceStatus.CONTINUE_ITERATING}
-                    ),
-                    ConvergenceStatusCollection(
-                        {"div_crit": ConvergenceStatus.CONTINUE_ITERATING}
-                    ),
-                    ConvergenceInfoCollection({"crit": 1.0}),
-                )
-            if model.time_step_converged[model.time_step_idx] is True:
-                return (
-                    ConvergenceStatusCollection({"crit": ConvergenceStatus.CONVERGED}),
-                    ConvergenceStatusCollection(
-                        {"div_crit": ConvergenceStatus.CONTINUE_ITERATING}
-                    ),
-                    ConvergenceInfoCollection({"crit": 0.0}),
-                )
-            else:
-                return (
-                    ConvergenceStatusCollection(
-                        {"crit": ConvergenceStatus.CONTINUE_ITERATING}
-                    ),
-                    ConvergenceStatusCollection({"div_crit": ConvergenceStatus.FAILED}),
-                    ConvergenceInfoCollection({"crit": np.nan}),
-                )
-
-    model = DynamicTimeStepTestCaseModel(
-        num_nonlinear_iterations=num_nonlinear_iterations,
-        time_step_converged=time_step_converged,
-        params={
-            "time_manager": time_manager,
-            "times_to_export": [],  # Suspends export
-        },
-    )
-    solver_params = {
-        "nonlinear_solver": DynamicNewtonSolver,
-        "nl_convergence_inc_atol": 1e-6,
-        "nl_max_iterations": MAX_NONLINEAR_ITER,
-    }
-
-    status = pp.ModelRunner(model, solver_params).run()
-    assert np.allclose(model.time_step_history, exported_dt_expected)
-    assert model.time_manager.final_time_reached() != should_fail
-    if should_fail:
-        assert isinstance(status, ModelRunnerStatusFailure)
-        assert status.reason.find(failure_reason) != -1
-    else:
-        assert status.is_success()
